+++ src/Akeneo/Pim/Automation/RuleEngine/front/__tests__/src/pages/EditRules/EditRulesContent.test.tsx
@@ -2,8 +2,9 @@ import React from 'react';
 import { EditRulesContent } from '../../../../src/pages/EditRules/EditRulesContent';
 import userEvent from '@testing-library/user-event';
 import { wait } from '@testing-library/dom';
-import { render, act } from '../../../../test-utils';
+import { render, act, screen } from '../../../../test-utils';
 import { IndexedScopes } from '../../../../src/repositories/ScopeRepository';
+import { Security } from '../../../../src/dependenciesTools';
 
 const setIsDirty = (_isDirty: boolean) => {};
 
@@ -16,6 +17,7 @@ describe('EditRulesContent', () => {
       code: ruleDefinitionCode,
       type: 'product',
       priority: 0,
+      enabled: true,
       actions: [],
       conditions: [],
       labels: { en_US: 'toto' },
@@ -39,6 +41,7 @@ describe('EditRulesContent', () => {
         meta: {},
       },
     };
+    const security: Security = { isGranted: (_acl: string) => true };
     // When
     const { findByText, findByLabelText } = render(
       <EditRulesContent
@@ -47,6 +50,7 @@ describe('EditRulesContent', () => {
         locales={locales}
         scopes={scopes}
         setIsDirty={setIsDirty}
+        security={security}
       />,
       {
         legacy: true,
@@ -72,6 +76,7 @@ describe('EditRulesContent', () => {
       code: ruleDefinitionCode,
       type: 'product',
       priority: 0,
+      enabled: true,
       actions: [],
       conditions: [],
       labels: { fr_FR: 'toto' },
@@ -101,6 +106,7 @@ describe('EditRulesContent', () => {
         meta: {},
       },
     };
+    const security: Security = { isGranted: (_acl: string) => true };
     // When
     const { findByTestId, findByText, findByLabelText } = render(
       <EditRulesContent
@@ -109,6 +115,7 @@ describe('EditRulesContent', () => {
         locales={locales}
         scopes={scopes}
         setIsDirty={setIsDirty}
+        security={security}
       />,
       {
         legacy: true,
@@ -130,5 +137,131 @@ describe('EditRulesContent', () => {
     await act(() => userEvent.type(usLabelInput, 'The new label'));
     // Then
     await wait(() => expect(titleDiv).toHaveTextContent('The new label'));
+  });
+
+  it('does not display the "save and execute" button when user has not the permission', async () => {
+    // Given
+    const ruleDefinitionCode = 'my_code';
+    const ruleDefinition = {
+      id: 1,
+      code: ruleDefinitionCode,
+      type: 'product',
+      priority: 0,
+      enabled: true,
+      actions: [],
+      conditions: [],
+      labels: { fr_FR: 'toto' },
+    };
+    const locales = [
+      {
+        code: 'en_US',
+        label: 'English (United States)',
+        region: 'United States',
+        language: 'English',
+      },
+      {
+        code: 'fr_FR',
+        label: 'French (France)',
+        region: 'France',
+        language: 'French',
+      },
+    ];
+    const scopes: IndexedScopes = {
+      ecommerce: {
+        code: 'ecommerce',
+        currencies: ['EUR', 'USD'],
+        locales: [locales[0]],
+        category_tree: 'master',
+        conversion_units: [],
+        labels: { en_US: 'e-commerce' },
+        meta: {},
+      },
+    };
+    const security: Security = { isGranted: (_acl: string) => false };
+    // When
+    render(
+      <EditRulesContent
+        ruleDefinitionCode={ruleDefinitionCode}
+        ruleDefinition={ruleDefinition}
+        locales={locales}
+        scopes={scopes}
+        setIsDirty={setIsDirty}
+        security={security}
+      />,
+      {
+        legacy: true,
+      }
+    );
+    // Then
+    expect(
+      screen.queryByText('pimee_catalog_rule.form.edit.execute.button')
+    ).not.toBeInTheDocument();
+  });
+
+  it('displays or hides the "save and execute" button according to the status', async () => {
+    // Given
+    const ruleDefinitionCode = 'my_code';
+    const ruleDefinition = {
+      id: 1,
+      code: ruleDefinitionCode,
+      type: 'product',
+      priority: 0,
+      enabled: true,
+      actions: [],
+      conditions: [],
+      labels: { fr_FR: 'toto' },
+    };
+    const locales = [
+      {
+        code: 'en_US',
+        label: 'English (United States)',
+        region: 'United States',
+        language: 'English',
+      },
+      {
+        code: 'fr_FR',
+        label: 'French (France)',
+        region: 'France',
+        language: 'French',
+      },
+    ];
+    const scopes: IndexedScopes = {
+      ecommerce: {
+        code: 'ecommerce',
+        currencies: ['EUR', 'USD'],
+        locales: [locales[0]],
+        category_tree: 'master',
+        conversion_units: [],
+        labels: { en_US: 'e-commerce' },
+        meta: {},
+      },
+    };
+    const security: Security = { isGranted: (_acl: string) => true };
+    // When
+    render(
+      <EditRulesContent
+        ruleDefinitionCode={ruleDefinitionCode}
+        ruleDefinition={ruleDefinition}
+        locales={locales}
+        scopes={scopes}
+        setIsDirty={setIsDirty}
+        security={security}
+      />,
+      {
+        legacy: true,
+      }
+    );
+    // Then
+    expect(
+      await screen.findByText('pimee_catalog_rule.form.edit.execute.button')
+    ).toBeInTheDocument();
+    // When
+    await act(async () => {
+      userEvent.click(await screen.findByTestId('edit-rules-input-status'));
+    });
+    // Then
+    expect(
+      screen.queryByText('pimee_catalog_rule.form.edit.execute.button')
+    ).not.toBeInTheDocument();
   });
 });