+++ src/Akeneo/Pim/Automation/RuleEngine/front/src/models/RuleDefinition.ts
@@ -6,6 +6,7 @@ export type RuleDefinition = {
   code: string;
   labels: { [localeCode: string]: string };
   priority: number;
+  enabled: boolean;
   conditions: Condition[];
   actions: Action[];
 };