+++ src/Akeneo/Pim/Automation/RuleEngine/front/src/pages/EditRules/edit-rules.types.ts
@@ -1,6 +1,7 @@
 type FormData = {
   code: string;
   priority: string;
+  enabled: boolean;
   labels: {
     [key: string]: string;
   };